// src/app/models/roleUpdate.ts
export class RoleUpdate{
  name: string;
  price: number;
  age: number;
  growth: number;
  clothsize: number;
  chestsize: number;
  waistsize: number;
  thighsize: number;
  footsize: number;
  necksize: number;
  headsize: number;
  tatu: boolean;
  piercing: boolean;
  appearancetype :number;
  haircolor: number;
  city: number;
  eyecolor: number;
  gender: number;
  country: number;
  employmenttype: number;
}